[rename] response dto 잘못된 배치 수정
