+++ order/src/main/java/com/sparta/rooibus/order/application/dto/response/DeleteOrderResponseDTO.java
@@ -1,4 +1,4 @@
-package com.sparta.rooibus.order.application.dto.request;
+package com.sparta.rooibus.order.application.dto.response;
 
 import com.sparta.rooibus.order.domain.entity.Order;
 import java.util.UUID;
